requirements.txt
transformers>=4.30.0
torch>=2.0.0
sentencepiece>=0.1.99
flask>=2.3.0
flask-cors>=4.0.0
Pillow>=10.0.0
